requirements.txt
# Core dependencies
streamlit>=1.40.0
langgraph>=1.0.0
langchain-ollama>=0.2.0
langchain-groq>=0.2.0
tavily-python>=0.2.0
python-dotenv>=1.0.0

# PDF generation
fpdf2>=2.7.0
markdown>=3.5.0

# API dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0

# Development dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.14.0
ruff>=0.4.0
mypy>=1.10.0
